Add common base view for export modals
